CrowdfundPro/crowdfundpro_backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==25.1
python-decouple==3.8
dj-database-url==3.0.0
psycopg2-binary==2.9.10
stripe==7.8.0
Pillow==11.2.1
celery==5.3.4
redis==5.0.1
django-q==1.3.9
python-dotenv==1.0.0 